+++ src/menu/AnimationFadeToBlack.java
@@ -44,6 +44,6 @@ public class AnimationFadeToBlack extends Animation {
 	
 	public void draw(MainWindow m) {
 //		m.setTexture(state.getPathToAnims() + texture);
-		m.renderNonTextured(x,y,width,height,alpha);
+		m.renderNonTextured((int)x,(int)y,width,height,alpha);
 	}
 }
