+++ src/menu/BackButton.java
@@ -8,7 +8,7 @@ public class BackButton extends ButtonMenuItem {
 	public BackButton(StartupNew m) {
 		super("Back",0,0,100,50, m);
 		text = "Back";
-		textObject = new Text(true,text,x,y,0,0,state.charList);
+		textObject = new Text(true,text,(int)x,(int)y,0,0,state.charList);
 		textObject.setAsSingleString();
 	}
