+++ src/menu/TexturedMenuItem.java
@@ -14,7 +14,7 @@ public class TexturedMenuItem extends MenuItem {
 	private int borderDH;
 	private int hoveredDy;
 	
-	public TexturedMenuItem(String text, int x, int y, int width, int height,  StartupNew state, String texture, int dx, int dy, int dw, int dh) {
+	public TexturedMenuItem(String text, double x, double y, int width, int height,  StartupNew state, String texture, int dx, int dy, int dw, int dh) {
 		super(text,x,y,width,height,state);
 		this.texture = texture;
 		this.dx = dx;
